Move async API logic to separate page
