+++ Source/SelfService/Web/integrations/connection/routes.tsx
@@ -11,8 +11,9 @@ import { ConsumeDataRestAPIView } from './connectionDetails/consumeDataRestAPI';
 import { MessagesListView } from './connectionDetails/messages/messagesList';
 import { MessagesViewRoot } from './connectionDetails/messages';
 import { Index as ChangeMessageView } from './connectionDetails/messages/changeMessage';
-
 import { NewConnectionView } from './new';
+import { ConsumeDataEventStreamsView } from './connectionDetails/consumeDataEventStreams';
+
 
 export const routes: RouteObject[] = [
     {
@@ -44,6 +45,10 @@ export const routes: RouteObject[] = [
             {
                 path: 'consume-data-rest-api',
                 element: <ConsumeDataRestAPIView />,
+            },
+            {
+                path: 'consume-data-event-streams',
+                element: <ConsumeDataEventStreamsView />,
             },
         ],
     },
